% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{extract_parms}
\alias{extract_parms}
\title{Extract parameters.}
\usage{
extract_parms(
  parms,
  pfCycleLength = NA_real_,
  inflec = NA_real_,
  ring_duration = NA_real_,
  seq_upper = NA_real_
)
}
\description{
Extract parameters.
}
